requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
black==22.8.0
boto3==1.24.64
botocore==1.27.64
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
password-strength==0.0.3.post2
pathspec==0.10.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
